chore(linter-formatter): move path
