+++ packages/@textlint/linter-formatter/src/FormatterOptions.ts
